// server/src/config/interfaces/query-parser-config.interface.ts
export enum FieldType {
     STRING = 'string', 
     BOOLEAN = 'boolean',
     NUMBER = 'number'
}

interface TypeOrmField {
    path: string; // dot-separated path
    type?: FieldType; // type of that field
    isViewEntity?: boolean; // 
}

export interface QueryParserConfiguration {
    fields: Record<string, TypeOrmField | TypeOrmField[]>; // can be either one field, either multiple fields (in case if it's object)
    includeFields?: string[];
    orderFields?: string[];
    filterFields?: string[];
    enableFilterFallbackCollection: boolean;
    searchField?: string; // field slug that would be used for ?search=
}